Fix rom-converter.py crashing on Windows
On Windows, rom-converter.py crashes during the .sfc -> .smc conversion step on line 50 (`os.remove(filename`) with a `PermissionError: [WinError 32] The process cannot access the file because it is being used by another process`.

This is because on Windows, it's illegal to remove a file while its handle is still being used somewhere. The use occurs earlier in the script where `inFile` handle is acquired but it's never explicitly closed. Adding a `inFile.close()` to line 33 fixes this Windows issue.